delete project.d.ts and move interfaces in with source code
